nlp/requirements.txt
spacy==3.7.2
langdetect==1.0.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0